requirements.txt
streamlit
langchain
langchain_google_genai
faiss-cpu
PyPDF2
python-dotenv
google-generativeai
langchain_community